feat: 랜딩 페이지 아이콘 섹션 일부 구현
