+++ apps/web/src/app/(landing)/_components/IconBox/IconBox.css.ts
@@ -4,13 +4,19 @@ export const accentVar = createVar()
 
 const bounce = keyframes({
   '0%': {
+    transform: 'translateY(-80px)',
+  },
+  '25%': {
     transform: 'translateY(0)',
   },
   '50%': {
     transform: 'translateY(-80px)',
   },
-  '100%': {
+  '75%': {
     transform: 'translateY(0)',
+  },
+  '100%': {
+    transform: 'translateY(-80px)',
   },
 })
 
@@ -23,7 +29,7 @@ export const IconBoxContainer = style({
   justifyContent: 'center',
   alignItems: 'center',
   backgroundColor: 'inherit',
-  animation: `${bounce} 3s infinite`,
+  animation: `${bounce} 3s forwards`,
 })
 
 export const CardContainer = style({
