+++ apps/web/src/app/(landing)/_components/IconSectionTitle/index.ts
@@ -1,0 +1,1 @@
+export { IconSectionTitle } from './IconSectionTitle'
